🐛 Fix the type of the organization_id property in the user creation response
It should be an integer, but it was a string.
As this request could contain a file, it's sent via FormData, instead of JSON, so the organization_id was received as a string. This is how the HTTP requests works: everything is a string, except when using JSON.
The returned object was made from the bare fields, so the organization_id was returned as string.