♻️ 🌱 Refactor the seeders for better cohesion and responsibility
The OrganizationSeeder was the root of every other example seeders. Now they have their own classes, like UserSeeder and PostSeeder, which is way easier to use and maintain.
It was mandatory for the development of the comments seeder, or at least necessary to avoid to much complicated code.
That bad practice was not a choice, it just came from the lack of Laravel knowledge.