♻️ 🌱 Move the call of the RoleSeeder from the DatabaseSeeder to the corresponding migration
As reactions types are not example data, it's not really smart to generate them in the DatabaseSeeder, even with the condition to avoid it in production. Seeding all the database, with the option --seed or any other way, should only lead to example data generation, not mandatory ones.
The other mandatory data of the app, ReactionType, have already been developed this way in a previous commit.