♻️ Simplify the migrations which create the posts and comments tables
Due to a misunderstanding, the foreign() method was used in place of the foreignId() one when the name of a foreign key was not matching the {table}_id pattern, like with author_id referencing the users table.
As this method foreign() method doesn't create a column, this operation was done manually, which was adding useless complexity.