🚚 Simplify the Post model related routes
All those routes was prefixed by /organizations/:organizationId, but in most cases this information was not needed. It's only due to a misunderstanding of the REST API principles.
Those routes are now defined under a unique group in the definition file, which is way more readable. The next ones to be refactored will follow the same new organization.