🥅 Fix a missing error catch when creating or updating an organization
When choosing a name that have already been taken, an SQL error was thrown, instead of a validation error.
In the actual client app, where no update is available for the moment, this case was very rare... but possible. If an organization was created with the same name between the first check (the validate_organization route) and the actual creation: boom.