🔒️ Add a middleware to logout disabled users
As it's not possible to logout another user than the current one when sessions are stored in files with Laravel, this middleware intercepts any further request from disabled users to log them out.
To be more complete and user friendly (and secured in some ways), a web socket signal should be added to that procedure, to logout a user a soon as he/she is disabled, instead of waiting for his/her next request. But this first layer of security is necessary to handle the cases where the browser is closed.