⏪ Revert the changes of the 7617bb06 commit
Removing the stack trace of 403 errors so hardly was not a really good idea. In fact, despite what the comment and the commit message were supposing, an environment setting can actually hide the stack traces: APP_DEBUG. When set to false (which must be done in production), no stack trace of any kind is shown.
So now, 403 stack traces will appear when needed on development machines.