🦺 Add the token field in the validation rules of the StoreUserRequest class
In the case of a member invitation, the email and the organization ID should not be present in the request body, because these informations will be retrieved from Redis. On the other side, no token should be sent when an administrator is creating his/her account, as it's not an invitation and he/she provides directly his/her email.