🥅 Add a 410 error in the user creation method for when the invitation does not exist
Despite handling direct POST requests which contains an expired or non existant token, this condition prevents a very rare case of normal use, where a user can visit the page of the invitation just a few seconds before it expires. In such scenario, he/she would be able to fill and send the form, but it would have raised an ugly system error.