🔒️ Add the suppression of all invitations related to a single user when creating the account
In this MVP, there is no UI to manage the invitations which have been sent. As an easy workaround, there is no limit on the number of invitations that an administrator can send to a single user, to be able to resend a new one if the previous one expired or if the email had been lost.
But to avoid security issues, and to avoid wasting storage, all the invitations related to a user are now removed when the account is created, and not only the one which was used to access the form.