✨ Add a GET route to return the authenticated user in session
There was no way to get the session user data after authentication, so the client was relying on localStorage to persist the user between page reloads and after closing the tab or the browser. This new approach, where the client fetches the user from the server before loading the app, is more secure and reliable.
Note: the case where the user is not authenticated is considered as a normal scenario, so no error is thrown and the response is empty.