🐛 Fix a browser cache issue with profile pictures
The previous approach used the /users/{id}/profile-picture API route to fetch user profile pictures. It worked fine... until a user uploaded a new one. Since the URL didn't change, the browser kept serving the old picture in cache, leading the user to believe the upload didn't work.
Now, profile pictures URLs are direct links to the files, which filenames are unique. This ensures that when a picture is updated, the browser is forced to fetch the new image instead of using the cached version.
Note: this new approach requires setting up the symbolic link between the root /public directory and the /storage/app/public one with the "php artisan storage:link" command.