🐛 Fix very rare error when a disabled user logs out voluntarily
The LogoutDisabledUser middleware wraps all API routes to automatically terminate sessions of disabled users on their next request. While this system worked well, it caused an issue when a disabled user attempted to log out on its own: the usual 401 error was thrown, showing the corresponding error page on the client app.
Now, the logout route is excluded from the LogoutDisabledUser middleware, ensuring no 401 error is thrown in this case anymore, resulting in a smoother user experience.