🌱 Remove `ReactionTypeSeeder` and `RoleSeeder`, move logic to related migrations
Since the content of these tables is necessary for the app to function correctly, it doesn't belong in seeders. Seeders are intended for populating sample data, so having a `ReactionTypeSeeder` and `RoleSeeder` was confusing.
Their logic has been moved to the corresponding migrations, improving developer experience by reducing seeders responsibility.