🗃️ 💥 Remove Role and ReactionType UI names from database
Given this data cannot be modified through the UI by the end user, having localized (french) UI names directly in the database was confusing and counterproductive. So the `name` field of each of these tables have been removed, and the previous `tag` one has been renamed to `name` to simplify the schema.
The client app must be updated accordingly to reflect this change.