♻️ 💥 Simplify `UserResource` and `ReactionResource`
Since the previous commit, the `Role` and `ReactionType` models no longer includes a UI name. So the API resources that use them can now simply return their system name to the client app, instead of the full model.
This change reduces the size of HTTP responses, and also simplifies the client codebase.