⚰️ Remove dead code in `checkNameConflict` method of `OrganizationController`
Currently, the app does not allow updating an organization, but the `checkNameConflict` method still contained code that was meant to support this feature. In fact, the unused `PATCH` route and its associated `update` method were removed in the `a8f8fdf` commit in September 2023, but the content of the `checkNameConflict` method was overlooked.
Removing this dead code improves the developer experience by simplifying the method and making it less prone to confusion.