🐛 Fix unintended profile picture deletion in `update` method of `UserController`
When a user deletes their profile picture or uploads a new one, the previous picture is removed from the file system. However, this deletion occurred too early, before ensuring the update was successful. As a result, if an error occurred, the profile picture was always deleted, even if the user hadn’t attempted to change it.