🐛 Fix another unintended profile picture deletion in `update` method of `UserController`
To prevent orphaned files, a user's new profile picture is removed from the file system if an error occurs during the database update. However, this deletion was performed without any condition. As a result, the profile picture was always deleted, even if the user hadn’t attempted to change it in the first place.